+++ src/utils/geom.ts
@@ -56,17 +56,8 @@ export function getColliderFromDie(die: Dice): Mesh {
     return collider;
 }
 
-function getFacetsPerFaceForDie(die: Dice): number {
-    if (die === Dice.D10) {
-        return 2;
-    } else {
-        return 1;
-    }
-}
-
 export function getValueFromFace(die: Dice, face: number): number {
-    const value = Math.floor(face / getFacetsPerFaceForDie(die));
-    return FACE_VALUE_MAPPING[die][value];
+    return FACE_VALUE_MAPPING[die][face];
 }
 
 export function stringToDice(dice: string): Dice {
@@ -138,7 +129,7 @@ const FACE_VALUE_MAPPING: Record<Dice, number[]> = {
         10,
     ],
     [Dice.D20]: [6, 9, 16, 3, 19, 11, 14, 8, 17, 1, 4, 20, 10, 7, 13, 18, 2, 12, 15, 5],
-    [Dice.D100]: [30, 50, 90, 70, 10, 100, 80, 20, 60, 40],
+    [Dice.D100]: [50, 90, 10, 70, 30, 40, 100, 80, 20, 60, 50, 90, 10, 70, 30, 40, 100, 80, 20, 60],
 };
 
 const d10Custom = {