+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 -   Invert vector for d4 value detection
+-   Update d10/d100 value detection code
 
 ## [0.1.0] - 2021-03-28